Support conversions from signed and 128-bit types
gcc and clang both provide native __int128 (signed and unsigned) on
the host side.  This adds implicit conversions from those types.

Also, for the conversion of signed types at 64 bits or smaller, if a
value is negative, sign-extend it into the high half for congruity
with how C handles widening signed-to-unsigned conversion (see ISO C99
6.3.1, summarized in https://unspecified.wordpress.com/2011/08/08/integer-conversions-in-c/ ).

This change does have one arguable regression: the uint128_t
constructor no longer works with any user-defined classes that can be
converted to uint64_t, which was previously allowed.  (This happens
because we call std::numeric_limits<T>, which we do to avoid a warning
about comparisons with 0 on unsigned integer types.)  However, since
user-defined classes might be a different large type (such as a
bignum), those conversions might have silently lost information
anyway, so the error hopefully alerts users to write a custom
constructor.