Reduce ambiguity in the operators.
Previously, in some circumstances, uint128_t operators would be
considered when it was pointless.  For instance, if you had
"(123 + FOO)" where FOO is an enum, then the compiler would consider
the operator+ we had defined.  That would produce messages like so:

foo.cu:26:25: warning: ISO C++ says that these are ambiguous, even though the worst conversion for the first is better than the worst conversion for the second: [enabled by default]

This was particularly common when used in conjunction with cub.

This change moves most binary operators from being friend functions to
being operator methods, so they're only considered when the first
argument is known to be a uint128_t.  The two binary operators that
weren't trivially changed (operator<< and operator>>) were modified to
only be considered for arithmetic types.